Move parsers to separate module
